requirements.txt
blinker==1.9.0
certifi==2026.2.25
paho-mqtt==1.6.1
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
Werkzeug==3.1.8
gunicorn
